refactor(core): split @templar/core into types-only kernel + @templar/engine
Strip @templar/core to a zero-runtime, zero-dependency protocol specification
(types-only). Move agent factory (createTemplar), validation, and middleware
wrapper to the new @templar/engine package. Move channel-specific runtime
utilities (CapabilityGuard, ChannelRegistry, block-utils, type-guards,
config-hash) to @templar/channel-base.

This is Phase M1 of the Lego Architecture migration (Issue #107):
- @templar/core: types-only kernel (Layer 0)
- @templar/engine: agent factory + validation (Layer 1)
- @templar/channel-base: absorbs channel infrastructure (Layer 2)

All 3,167 tests pass across 27 packages. Clean break — no backward-compat
re-exports.
