+++ packages/channel-base/src/config-hash.ts
