+++ packages/engine/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@templar/engine",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "scripts": {
+    "build": "tsup",
+    "typecheck": "tsc --noEmit",
+    "lint": "biome check .",
+    "test": "vitest run"
+  },
+  "dependencies": {
+    "@templar/core": "workspace:*",
+    "@templar/errors": "workspace:*"
+  },
+  "peerDependencies": {
+    "@langchain/langgraph": "^0.1.0",
+    "deepagents": "^1.0.0"
+  },
+  "peerDependenciesMeta": {
+    "@langchain/langgraph-supervisor": {
+      "optional": true
+    },
+    "@langchain/langgraph-swarm": {
+      "optional": true
+    }
+  },
+  "devDependencies": {
+    "@nexus/sdk": "workspace:*",
+    "@types/node": "^25.2.2"
+  }
+}
