+++ packages/engine/src/index.ts
@@ -1,0 +1,3 @@
+export { createTemplar } from "./create-templar.js";
+export { registerMiddlewareWrapper, unregisterMiddlewareWrapper } from "./middleware-wrapper.js";
+export { validateAgentType, validateManifest, validateNexusClient } from "./validation.js";
